For the "Automatic Iteration Notification" (#6) implemented the Break methos
When  "Automatic Iteration Notification"  feature is activated we cannot rely anymore on the possibility to stop internally the iterator (which is effective if the "Iteration Result" stop condition is enabled) since the "Notify Iteration Msg" is transparently sent. Indeed when the option is not activated sending this message allowed the user to communicate the iteration result by means of a Boolean flag, which when false would indicate the Internal stop.

For this reason we implemented the Break method and corresponding message to allow the user to stop the iterator internally and yield the appropriate stop reason. Nothing prevents the user to send the "Stop Iterations Msg",  when the "Automatic Iteration Notification" is enabled allowing the user to differentiate between an internal or external stop by means of the Break method helps the user to differentiate when the Iterator was stopped externally or internally.