Renamed Examples, Error generation, "Auto Iteration Notification" on. VIPM Package
We have renamed the examples so that  through the VIPM package we install the one with shortest name only, i.e. the one covering the Iterator Actor NR configured with "Auto Iteration Notification" enabled

Now generate an error to the caller to the caller if Break, Start Iterations, and Stop Iterations are called inappropriately.

"Auto Iteration Notification" on by default since this is the suggested way of operation

First Draft of a VIPM package